move information-dialog
