+++ src/app/shared/components/dialog-switcher-button/dialog-switcher-button.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { DialogSwitcherButtonComponent } from './dialog-switcher-button.component';
 import { MatDialogModule, MatDialog } from '@angular/material/dialog';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { InformationDialogComponent } from '../information-dialog/information-dialog.component';
+import { InformationDialogComponent } from '../components/information-dialog/information-dialog.component';
 
 describe('DialogSwitcherButtonComponent', () => {
   let component: DialogSwitcherButtonComponent;
