+++ src/app/shared/dialogs/information-dialog/information-dialog.component.css
