+++ src/app/shared/dialogs/information-dialog/information-dialog.component.ts
